replaced read service dependency
