+++ src/Submissions/Api/Services/Directory/BiobankReadService.cs
@@ -317,16 +317,6 @@ namespace Biobanks.Submissions.Api.Services.Directory
             false,
              x => x.ServiceOfferingId == id)).Count();
 
-        public async Task<IEnumerable<ApplicationUser>> ListBiobankAdminsAsync(int biobankId)
-        {
-            var adminIds = (await _organisationUserRepository.ListAsync(
-                false,
-                x => x.OrganisationId == biobankId))
-                .Select(x => x.OrganisationUserId);
-
-            return _userManager.Users.AsNoTracking().Where(x => adminIds.Contains(x.Id));
-        }
-
         public async Task<IEnumerable<ApplicationUser>> ListSoleBiobankAdminIdsAsync(int biobankId)
         {
             // Returns users who have admin role only for this biobank
