+++ src/Submissions/Api/Services/Directory/Contracts/IBiobankService.cs
@@ -1,4 +1,5 @@
-﻿using Biobanks.Entities.Data;
+using Biobanks.Data.Entities;
+using Biobanks.Entities.Data;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -7,6 +8,7 @@ namespace Biobanks.Submissions.Api.Services.Directory.Contracts
     public interface IBiobankService
     {
         Task<IEnumerable<OrganisationServiceOffering>> ListBiobankServiceOfferingsAsync(int biobankId);
+        Task<IEnumerable<ApplicationUser>> ListBiobankAdminsAsync(int biobankId);
 
-    }
+  }
 }
